+++ lib/internal_imports.dart
@@ -1,23 +1,30 @@
 // 📁 Domain - Repositories
 export 'src/features/pagination/domain/repositories/pagy_repository.dart';
+export 'src/features/pagination/domain/repositories/pagy_page_repository.dart';
 
 // 📁 Domain - Usecases
 export 'src/features/pagination/domain/usecases/get_paginated_data_usecase.dart';
+export 'src/features/pagination/domain/usecases/get_paginated_page_usecase.dart';
 
 // 📁 Data - Datasources
 export 'src/features/pagination/data/datasources/network_api_service.dart';
 
 // 📁 Data - Models
 export 'src/features/pagination/data/models/pagy_model.dart';
+export 'package:pagy/src/features/pagination/param/pagy_params.dart';
+
 export 'src/core/utils/pagy_utils.dart';
 
 // 📁 Data - Repositories Implementation
 export 'src/features/pagination/data/repositories/pagy_repository_impl.dart';
+export 'src/features/pagination/data/repositories/pagy_page_repository_impl.dart';
 
 export 'src/features/pagination/domain/entities/pagy_response_parser.dart';
 export 'src/features/pagination/domain/entities/pagy_state.dart';
+export 'src/features/pagination/domain/entities/pagy_page.dart';
 export 'src/features/pagination/domain/enums/pagy_enum.dart';
 export 'src/features/pagination/data/datasources/pagy_remote_datasource.dart';
+export 'src/features/pagination/param/pagy_page_params.dart';
 export 'src/features/pagination/presentation/widgets/common/pagy_empty_state_widget.dart';
 export 'src/features/pagination/presentation/widgets/common/pagy_error_widget.dart';
 export 'src/features/pagination/presentation/widgets/common/pagy_loading_widget.dart';